+++ client/src/context/search.jsx
@@ -1,4 +1,4 @@
-import { useState, createContext, useContext } from "react";
+import { useState, createContext, useContext, useEffect } from "react";
 
 const SearchContext = createContext();
 
@@ -13,13 +13,44 @@ const innitialState = {
   loading: false,
 };
 
-const SearchProvider = ({ children }) => {
+/*const SearchProvider = ({ children }) => {
   const [search, setSearch] = useState(innitialState);
 
-  console.log("SearchProvider render", { search, innitialState });
+  //console.log("SearchProvider render", { search, innitialState });
+
+  useEffect(() => {
+   // console.log("Search state updated:", search);
+  }, [search]); // 🔥 Tulostaa vain kun `search` muuttuu, ei ääretöntä silmukkaa!
 
   return (
     <SearchContext.Provider value={[search, setSearch, innitialState]}>
+      {children}
+    </SearchContext.Provider>
+  );
+};*/
+
+const SearchProvider = ({ children }) => {
+  const [search, setSearch] = useState(innitialState);
+
+  const handleSearch = async (query) => {
+    try {
+      const response = await axios.get(`/api/search`, {
+        params: query,
+      });
+      setSearch((prev) => ({
+        ...prev,
+        results: response.data,
+        page: '/search',
+        loading: false,
+      }));
+    } catch (error) {
+      console.error("Search error:", error);
+      setSearch((prev) => ({ ...prev, loading: false }));
+    }
+  };
+
+  return (
+    <SearchContext.Provider value={[search, setSearch, handleSearch]}>
       {children}
     </SearchContext.Provider>
   );
