+++ server/package.json
@@ -10,8 +10,9 @@
   "license": "ISC",
   "description": "",
   "dependencies": {
+    "@aws-sdk/client-s3": "^3.741.0",
+    "@aws-sdk/client-ses": "^3.741.0",
     "@eslint/config-array": "^0.18.0",
-    "aws-sdk": "^2.16.0",
     "axios": "^1.7.7",
     "bcrypt": "^5.1.1",
     "cors": "^2.8.5",
