+++ server/server.js
@@ -1,81 +1,103 @@
-import 'dotenv/config';
+import dotenv from 'dotenv';
+import 'dotenv/config'; 
 import express from "express";
+import geocodeRoutes from "./routes/geocode.js";
 import morgan from "morgan";
 import cors from "cors";
 import mongoose from "mongoose";
 import { DATABASE } from "./config.js";
 import authRoutes from "./routes/auth.js";
 import adRoutes from "./routes/ad.js";
-import axios from "axios"; // Axios tarvitaan Geocoding API -kutsuja varten
+import axios from "axios";
 
+dotenv.config();
 const app = express();
-app.use(cors());
+
+// CORS-asetukset oikein
+/*app.use(cors({
+  origin: ['http://localhost:5173', 'https://real-estate0824.onrender.com'],
+  methods: ["GET", "POST"],
+  allowedHeaders: ["Content-Type", "Authorization"]
+}));*/
+
+
+/*app.use(cors({
+    origin: 'http://localhost:5173',  // Salli frontendin osoite
+    credentials: true,
+    allowedHeaders: ['Content-Type', 'Authorization', 'refresh_token']  // Lisää refresh_token
+  }));*/
+
+  app.use(cors({
+    origin: ['http://localhost:5173', 'http://localhost:5174'], // Lisää kaikki sallitut originit
+    credentials: true,
+    allowedHeaders: ['Content-Type', 'Authorization', 'refresh_token']
+}));
+
+app.options("*", cors());
 app.use(express.static('dist'));
 
-// db
-mongoose.set("strictQuery", false);
+// DB-yhteys
+/*mongoose.set("strictQuery", false);
 mongoose
   .connect(DATABASE)
   .then(() => console.log("DB connected"))
-  .catch((err) => {
-    console.error("DB connection error:", err.message);
-  });
+  .catch((err) => console.error("DB connection error:", err.message));*/
 
-// middlewares
+  const connectDB = async () => {
+    try {
+      await mongoose.connect(DATABASE); 
+      console.log("✅ MongoDB Connected");
+    } catch (error) {
+      console.error("❌ MongoDB connection failed:", error);
+      process.exit(1); // Sulkee sovelluksen, jos tietokantayhteys epäonnistuu
+    }
+  };
+  
+  // Yhdistetään tietokantaan ennen kuin server käynnistyy
+  connectDB();
+
+// Middlewares
 app.use(express.json({ limit: "10mb" }));
 app.use(morgan("dev"));
+//app.use(cors());
 
-app.use(cors({
-  origin: ['http://localhost:5173', 'https://real-estate0824.onrender.com']
-}));
-
-
-// routes middleware
+// Reitit
 app.use("/api", authRoutes);
 app.use("/api", adRoutes);
+app.use("/api", geocodeRoutes);
 
-// Geocoding API -reitti
-/*const router = express.Router();
+console.log("MongoDB URI:", process.env.DATABASE);
+console.log("Google API Key from ENV:", process.env.GOOGLE_GEOCODE_API_KEY);
 
-router.get('/geocode', async (req, res) => {
-  console.log("Request received for /geocode");
-  console.log("Query parameters:", req.query);
-  const { address } = req.query;
+// Geocoding API -reitti Expressin kautta
+/*app.get('/geocode', async (req, res) => {
+  try {  
+    const { place_id } = req.query;
+    console.log("Received place_id:", place_id); // Lisätty lokitus
+    if (!place_id) return res.status(400).json({ error: "place_id is required" });
 
-  if (!address) {
-    return res.status(400).json({ error: "Address is required" });
-  }
-
-  try {
-    const response = await axios.get('https://maps.googleapis.com/maps/api/geocode/json', {
-      params: {
-        address,
-        key: process.env.GOOGLE_API_KEY,
-      },
+   // console.log("Using Google API Key:", apiKey); // Lisätty lokitus
+    const apiKey = process.env.GOOGLE_GEOCODE_API_KEY;
+    const response = await axios.get(`https://maps.googleapis.com/maps/api/geocode/json`, {
+      params: { place_id, key: apiKey }
     });
 
-    if (response.data.status !== "OK") {
-      return res.status(400).json({ error: response.data.error_message || "Geocoding failed" });
-    }
+    console.log("Google Geocode API response:", response.data); // Lisätty lokitus
 
     res.json(response.data);
-  } catch (err) {
-    console.error("Geocoding API error:", err.message);
-    res.status(500).json({ error: "Internal server error" });
+  } catch (error) {
+    console.error("Geocoding error:", error.message);
+    res.status(500).json({ error: "Google Geocoding request failed" });
   }
-});
+});*/
 
-app.use('/api', router);*/
-
-// Testi reitti palvelimessa
+// Testi reitti
 app.get("/test", (req, res) => {
   res.send("Test route works!");
 });
 
+// Palvelimen käynnistys
 const PORT = process.env.PORT || 8000;
 app.listen(PORT, () => {
   console.log(`Server is running on port ${PORT}`);
 });
-
-/*netstat -ano | findstr 8000
-taskkill /F /PID */
